hostel_management/requirements.txt
asgiref==3.7.2
Django==4.1.7
django-cors-headers==4.0.0
django-dotenv==1.4.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
celery==5.1.2
redis==3.5.3
pandas==2.0.2
gunicorn==20.1.0